chore: restructured types
